feat: desktop as default route, setup wizard with security & telegram steps
- Move desktop UI to / as the default landing page, setup wizard at /setup
- Add Security (password/hotspot) and Telegram steps to setup wizard (6 steps total)
- Deduplicate getGatewayToken() into shared gateway-proxy module
- Update Chromium icon to proper Chromium SVG
- Parallelize browser lock file cleanup
- Move system update button from setup to settings app
- Remove stats from DoneStep (already in desktop UI)
