+++ next.config.ts
@@ -15,7 +15,7 @@ const APP_VERSION = (() => {
 const nextConfig: NextConfig = {
   output: "standalone",
   serverExternalPackages: ["better-sqlite3"],
-  allowedDevOrigins: ["http://10.69.0.9:3005"],
+  allowedDevOrigins: ["http://clawbox.local"],
   devIndicators: false,
   compress: true,
   poweredByHeader: false,
