+++ src/components/SystemTray.tsx
@@ -1,6 +1,8 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useState, useEffect, useCallback, useRef } from "react";
+
+const BRAND_ORANGE = "#fe6e00";
 
 interface SystemTrayProps {
   isOpen: boolean;
@@ -8,6 +10,13 @@ interface SystemTrayProps {
   date: string;
   time: string;
 }
+
+type RebootState =
+  | null
+  | { phase: "waiting"; action: "restart" | "shutdown"; dots: number }
+  | { phase: "reconnecting"; attempt: number; dots: number }
+  | { phase: "restoring" }
+  | { phase: "shutdown" };
 
 export default function SystemTray({
   isOpen,
@@ -17,6 +26,9 @@ export default function SystemTray({
 }: SystemTrayProps) {
   const [closing, setClosing] = useState(false);
   const [confirmAction, setConfirmAction] = useState<"shutdown" | "restart" | null>(null);
+  const [rebootState, setRebootState] = useState<RebootState>(null);
+  const pollRef = useRef<ReturnType<typeof setInterval> | null>(null);
+  const dotsRef = useRef<ReturnType<typeof setInterval> | null>(null);
 
   useEffect(() => {
     if (isOpen) {
@@ -25,7 +37,59 @@ export default function SystemTray({
     }
   }, [isOpen]);
 
+  // Animate dots
+  useEffect(() => {
+    if (rebootState && rebootState.phase !== "restoring") {
+      dotsRef.current = setInterval(() => {
+        setRebootState(prev => {
+          if (!prev || prev.phase === "restoring" || prev.phase === "shutdown") return prev;
+          return { ...prev, dots: (prev.dots + 1) % 4 };
+        });
+      }, 500);
+      return () => { if (dotsRef.current) clearInterval(dotsRef.current); };
+    }
+  }, [rebootState?.phase]);
+
+  const startReconnecting = useCallback(() => {
+    setRebootState({ phase: "reconnecting", attempt: 0, dots: 0 });
+
+    let attempt = 0;
+    pollRef.current = setInterval(async () => {
+      attempt++;
+      setRebootState(prev => {
+        if (!prev || prev.phase !== "reconnecting") return prev;
+        return { ...prev, attempt };
+      });
+
+      try {
+        const res = await fetch("/setup-api/setup/status", { signal: AbortSignal.timeout(3000) });
+        if (res.ok) {
+          // Device is back online
+          if (pollRef.current) clearInterval(pollRef.current);
+          setRebootState({ phase: "restoring" });
+          setTimeout(() => {
+            setRebootState(null);
+            window.location.reload();
+          }, 1500);
+        }
+      } catch {
+        // Still offline, keep polling
+      }
+    }, 3000);
+
+    return () => { if (pollRef.current) clearInterval(pollRef.current); };
+  }, []);
+
+  // Cleanup on unmount
+  useEffect(() => {
+    return () => {
+      if (pollRef.current) clearInterval(pollRef.current);
+      if (dotsRef.current) clearInterval(dotsRef.current);
+    };
+  }, []);
+
   const handleClose = () => {
+    if (rebootState) return; // Don't close during reboot
     setClosing(true);
     setTimeout(onClose, 150);
   };
@@ -35,6 +99,10 @@ export default function SystemTray({
       setConfirmAction(action);
       return;
     }
+
+    // Show the waiting overlay
+    setRebootState({ phase: "waiting", action, dots: 0 });
+
     try {
       await fetch("/setup-api/system/power", {
         method: "POST",
@@ -42,8 +110,93 @@ export default function SystemTray({
         body: JSON.stringify({ action }),
       });
     } catch {}
-    handleClose();
+
+    if (action === "restart") {
+      // Wait for the device to go down, then start reconnecting
+      setTimeout(() => startReconnecting(), 5000);
+    } else {
+      // Shutdown: show spinner briefly, then show powered-off screen
+      setTimeout(() => setRebootState({ phase: "shutdown" }), 4000);
+    }
   };
+
+  // Full-screen reboot overlay
+  if (rebootState) {
+    const dots = (rebootState.phase === "waiting" || rebootState.phase === "reconnecting")
+      ? ".".repeat(rebootState.dots)
+      : "";
+
+    return (
+      <div className="fixed inset-0 z-[99999] flex items-center justify-center" style={{ background: rebootState.phase === "shutdown" ? "rgba(0, 0, 0, 0.97)" : "rgba(0, 0, 0, 0.92)" }}>
+        <div className="flex flex-col items-center gap-6 max-w-md text-center px-6">
+          {/* Icon */}
+          {rebootState.phase === "restoring" ? (
+            <div className="w-16 h-16 rounded-full flex items-center justify-center" style={{ backgroundColor: BRAND_ORANGE }}>
+              <span className="material-symbols-rounded text-white" style={{ fontSize: 32 }}>check</span>
+            </div>
+          ) : rebootState.phase === "shutdown" ? (
+            <div className="w-16 h-16 rounded-full flex items-center justify-center bg-white/10">
+              <span className="material-symbols-rounded text-white/60" style={{ fontSize: 32 }}>power_settings_new</span>
+            </div>
+          ) : (
+            <div className="w-16 h-16 rounded-full border-[3px] border-white/10 animate-spin" style={{ borderTopColor: BRAND_ORANGE }} />
+          )}
+
+          {/* Status message */}
+          {rebootState.phase === "waiting" && (
+            <>
+              <h2 className="text-xl font-semibold text-white">
+                {rebootState.action === "shutdown" ? "Shutting down" : "Restarting"}{dots}
+              </h2>
+              <p className="text-sm text-white/50">
+                {rebootState.action === "shutdown"
+                  ? "Your device is shutting down. Please wait."
+                  : "Please wait while your device restarts. This may take a minute."}
+              </p>
+            </>
+          )}
+
+          {rebootState.phase === "shutdown" && (
+            <>
+              <h2 className="text-xl font-semibold text-white">
+                ClawBox has been shut down
+              </h2>
+              <p className="text-sm text-white/50">
+                Your device has been powered off.
+              </p>
+              <div className="mt-2 px-4 py-3 rounded-lg bg-white/5 border border-white/10">
+                <p className="text-sm text-white/70">
+                  To start your ClawBox again, unplug the power cable and plug it back in.
+                </p>
+              </div>
+            </>
+          )}
+
+          {rebootState.phase === "reconnecting" && (
+            <>
+              <h2 className="text-xl font-semibold text-white">
+                Reconnecting
+              </h2>
+              <p className="text-sm text-white/50">
+                Waiting for device to come back online
+              </p>
+            </>
+          )}
+
+          {rebootState.phase === "restoring" && (
+            <>
+              <h2 className="text-xl font-semibold text-white">
+                Device is back online
+              </h2>
+              <p className="text-sm text-white/50">
+                Restoring your session...
+              </p>
+            </>
+          )}
+        </div>
+      </div>
+    );
+  }
 
   if (!isOpen && !closing) return null;
